requirements.txt
dj-database-url==0.5.0
Django==2.2
gunicorn==19.9.0
psycopg2-binary==2.7.7
whitenoise==4.1.2
python-decouple==3.1
django-hosts==3.0
Pillow==5.4.1
django-imagekit==4.0.2
numpy
django-subdomains==2.1.0
